Fix: regex fails to capture certain langs
Languages with non alphanumeric characters (i.e. c++)
would not be captured by the regex for syntax highlighting.

This change moves from a positive capture (all alphanumeric
characters) to a negative capture (anything that is NOT a ").

This should be a more robust solution... at least until some-
one invents a language with a " in the name.